amf: deserialise AMF request
Extract common code that deserialises the AMF message to an abstract
class, AbstractAMFTextView.
Add a new class, RequestAMFTextView, to (deserialise and) show the AMF
request.
Change class ResponseAMFTextView to extend AbstractAMFTextView.
Change ExtensionAMF to install/uninstall the request view and initialise
common code.
Bump version and update changes in ZapAddOn.xml file.
