+++ src/org/zaproxy/zap/extension/amf/ZapAddOn.xml
@@ -1,13 +1,13 @@
 <zapaddon>
     <name>AMF</name>
-    <version>1</version>
+    <version>2</version>
     <status>alpha</status>
     <description>Adds support for AMF messages</description>
     <author>ZAP Dev Team</author>
     <url/>
     <changes>
     <![CDATA[
-    First release as an add-on, previously bundled with core.
+    Deserialise the AMF request.
     ]]>
     </changes>
     <dependson/>
